requirements.txt
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.1.2
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pycparser==2.21
python-slugify==6.1.1
six==1.16.0
slugify==0.0.1
SQLAlchemy==1.4.35
text-unidecode==1.3
uni-slugify==0.1.4
Unidecode==1.3.4
Werkzeug==2.1.1